requirements.txt
aiohttp==3.9.0rc0
aiosignal==1.3.1
attrs==23.1.0
discord-py-interactions==5.13.1
discord-typings==0.9.0
dnspython==2.4.2
emoji==2.8.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
pymongo==4.6.0
tomli==2.0.1
typing_extensions==4.8.0
yarl==1.9.2
gspread==6.1.2
